refactor: Initial, Result 통합
